Refactoring / organization.
