+++ core/testing/edu/asu/c3simulator/testing/stubs/PlayerTestingStub.java
@@ -1,4 +1,4 @@
-package edu.asu.c3simulator.testing;
+package edu.asu.c3simulator.testing.stubs;
 
 import java.text.DateFormat;
