+++ core/testing/edu/asu/retired/testing/automated/TestWidget.java
@@ -1,4 +1,4 @@
-package edu.asu.c3simulator.testing.automated;
+package edu.asu.retired.testing.automated;
 
 import static org.junit.Assert.assertTrue;
